fix: really attempt to return _all_ bytes if plain read called
The previous behaviour limits a single read to 2^63 - 1 bytes, even if all are
requested. However, this is wrong. If all are requested all should be returned.

There is a (strong!) possibility that more bytes than 2^63 - 1 cannot be
allocated. But then this is a reason that the function should error - it should
not behave as though 2^63 - 1 bytes is the entire stream.

This is probably all academic... I suspect there isn't a system around running
Python that could come close to allocating 2^63 - 1 bytes.